backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-docx==1.1.2
openpyxl==3.1.2
httpx==0.27.0
pydantic==2.7.4
pytest==8.3.2
